feat: Implement check-in functionality for daily attendance
- Added new component for daily check-in including date selection and student selection.
- Created HTML and SCSS files for the new check-in component.
- Updated utility functions to ensure proper date formatting.
- Refactored existing check-in component to streamline the process and improve user experience.
- Integrated auto-complete for student selection and calendar for date selection.
- Implemented form validation and error handling for user inputs.
