+++ src/app/app.routes.ts
@@ -213,6 +213,7 @@ import { FrequenciaDashboardComponent } from './trilhar/frequencia/frequencia-da
 import { FrequenciaCheckinDiaComponent } from './trilhar/frequencia/frequencia-checkin-dia/frequencia-checkin-dia.component';
 import { FrequenciaGrupoComponent } from './trilhar/frequencia/frequencia-grupo/frequencia-grupo.component';
 import { FrequenciaCheckinDiaIncluirComponent } from './trilhar/frequencia/frequencia-checkin-dia-incluir/frequencia-checkin-dia-incluir.component';
+import { FrequenciaCheckinDiaIncluirDataComponent } from './trilhar/frequencia/frequencia-checkin-dia-incluir-data/frequencia-checkin-dia-incluir-data.component';
 
 export const routes: Routes = [
     // {path: '', component: EcommerceComponent},
@@ -260,10 +261,10 @@ export const routes: Routes = [
         children: [
           {path: 'dashboard', component: FrequenciaDashboardComponent, data: { rota: "dashboard", titulo: "Dashboard", breadcrumb: "Dashboard", } },
           {path: 'grupo', component: FrequenciaGrupoComponent, data: { rota: "grupo", titulo: "Resultado frequência por grupo", breadcrumb: "frequências por grupo", },},
-          {path: 'checkin-dia', component: FrequenciaCheckinDiaComponent, data: { rota: "checkin-dia", titulo: "Resultado checkin no dia", breadcrumb: "Checkin dia", },},
+          {path: 'checkin-dia', component: FrequenciaCheckinDiaComponent, data: { rota: "checkin-dia", titulo: "Resultado Check-in no dia", breadcrumb: "Checkin dia", },},
 
-          {path: 'checkin-dia/incluir', component: FrequenciaCheckinDiaIncluirComponent, data: { rota: "incluir", titulo: "Checkin no dia", breadcrumb: "Incluir", },},
-          // {path: 'checkin-dia/incluir/:data', component: FrequenciaCheckinDiaIncluirComponent, data: { rota: "incluir", titulo: "Checkin no dia", breadcrumb: "Alterar", },},
+          {path: 'checkin-dia/incluir', component: FrequenciaCheckinDiaIncluirComponent, data: { rota: "incluir", titulo: "Check-in no dia", breadcrumb: "Incluir", },},
+          {path: 'checkin-dia/incluir/:data', component: FrequenciaCheckinDiaIncluirDataComponent, data: { rota: "incluir", titulo: "Check-in no dia", breadcrumb: "Incluir", },},
       ]
     },
     {
