+++ src/app/trilhar/frequencia/frequencia-checkin-dia-incluir-data/frequencia-checkin-dia-incluir-data.component.html
@@ -1,0 +1,98 @@
+<!-- Breadcrumb -->
+<div class="breadcrumb-card mb-25 d-md-flex align-items-center justify-content-between">
+  <h5 class="mb-0">
+    {{this.getTitle()}} - data: {{ dataSelecionada | date:'dd/MM/yyyy' }}
+  </h5>
+  <ol class="breadcrumb list-unstyled mt-0 mb-0 pl-0">
+    <li class="breadcrumb-item position-relative">
+      <a routerLink="/" class="d-inline-block position-relative">
+        <i class="ri-home-8-line"></i>
+        Dashboard
+      </a>
+    </li>
+    <!-- <li class="breadcrumb-item position-relative">
+          <a [routerLink]="this.getRotaOperacao()" class="d-inline-block position-relative">
+              <i class="ri-arrow-left-line"></i>
+              {{this.getTitle()}}
+          </a>
+      </li> -->
+    <li class="breadcrumb-item position-relative">
+      {{this.getOperacao()}}
+    </li>
+  </ol>
+</div>
+
+<form [formGroup]="formularioCheckin">
+  <mat-card class="daxa-card events-card mb-25 border-radius bg-white border-none d-block"
+    [class.rtl-enabled]="themeService.isRTLEnabled()">
+
+    <mat-card-content>
+      <div class="row align-items-end">
+        <div class="col-xl-3 col-lg-3 col-md-3 col-sm-6 form-group" style="border: 0px solid red">
+          <div class="mb-25">
+            <app-calendario ngSkipHydration [formGroup]="formularioCheckin" [field]="'data'"
+              [label]="'Data do Check-in'" [maxDate]="maxData" (itemSelecionado)="onDataSelecionada($event)" [readonly]="true" >
+            </app-calendario>
+          </div>
+        </div>
+      </div>
+      <div class="row align-items-end">
+        <div class="col-xl-8 col-lg-8 col-md-8 col-sm-6 col-xs-6 form-group" style="border: 0px solid red">
+          <div class="mb-25">
+            <app-auto-complete ngSkipHydration [formGroup]="formularioCheckin" [field]="'alunoSelecionado'"
+              [itens]="listaAlunosAutoComplete" [propriedadeMostrar]="'descricao'" [label]="'Selecione uma criança'"
+              (itemSelecionado)="onListaAlunosAutoCompleteSelecionado($event)">
+            </app-auto-complete>
+          </div>
+        </div>
+        <div class="col-xl-4 col-lg-4 col-md-4 col-sm-6 col-xs-6 form-group" style="border: 0px solid red">
+          <div class="mb-25">
+            <button mat-flat-button color="accent" class="search-btn w-100" (click)="onRegistrarFrequencias()">
+              <mat-icon>event_available</mat-icon>Registrar Check-in
+            </button>
+          </div>
+        </div>
+      </div>
+
+      <div class="row">
+        <div class="col-xl-12 col-lg-12 col-md-12 col-sm-12 col-xs-12 form-group">
+
+          <!-- <div class="mb-20" *ngIf="alunoAtual?.matricula.turmaDescricao">
+            <h2>
+              <b>Sala:</b>
+              <span [ngStyle]="{ color: getCorSala(alunoAtual?.matricula.turmaDescricao) }">
+                {{ alunoAtual?.matricula.turmaDescricao }}
+              </span>
+            </h2>
+          </div> -->
+
+          <div class="mb-20" *ngIf="alunoAtual?.matricula.turmaDescricao">
+            <h2>
+              <b>Sala:</b> {{ alunoAtual?.matricula.turmaDescricao }}
+              <span class="badge"
+                [ngStyle]="{ backgroundColor: getCorSala(alunoAtual?.matricula.turmaDescricao), color: 'white', padding: '4px 8px', borderRadius: '6px', fontSize: '0.8rem' }">
+                {{ alunoAtual?.matricula.turmaDescricao.split(' ')[0] }}
+              </span>
+            </h2>
+          </div>
+
+
+          <div class="mb-20" *ngIf="alunoAtual?.nomeMae">
+            <h2><b>Mãe:</b> {{ alunoAtual?.nomeMae }}</h2>
+          </div>
+          <div class="mb-20" *ngIf="alunoAtual?.nomePai">
+            <h2><b>Pai:</b> {{ alunoAtual?.nomePai }}</h2>
+          </div>
+          <div class="mb-20" *ngIf="alunoAtual?.outroResponsavel">
+            <h2><b>Responsável:</b> {{ alunoAtual?.outroResponsavel }}</h2>
+          </div>
+          <div class="mb-20" *ngIf="alunoAtual?.dataNascimento">
+            <h2><b>Nascimento/Idade:</b> {{ alunoAtual?.idadeCrianca }}</h2>
+          </div>
+
+        </div>
+      </div>
+
+    </mat-card-content>
+  </mat-card>
+</form>
