+++ src/app/trilhar/frequencia/frequencia-checkin-dia-incluir-data/frequencia-checkin-dia-incluir-data.component.scss
@@ -1,0 +1,58 @@
+/* Adicione ao seu arquivo de estilos */
+.mat-table {
+  width: 100%;
+  border-collapse: collapse;
+}
+
+.mat-table th {
+  text-align: left;
+  padding: 6px;
+  font-weight: 500;
+  color: rgba(0, 0, 0, 0.87);
+  background-color: #f5f5f5;
+}
+
+.mat-table td {
+  padding: 6px;
+  border-bottom: 1px solid rgba(0, 0, 0, 0.12);
+}
+
+.mat-table tr:hover {
+  background-color: rgba(0, 0, 0, 0.04);
+}
+
+
+
+.search-box {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+}
+
+.search-input-wrapper {
+  flex: 1;
+}
+
+.search-button-wrapper {
+  flex-shrink: 0;
+}
+
+
+
+
+.link-turma {
+  display: inline-flex;
+  align-items: center;
+  cursor: pointer;
+  color: #3f51b5; // cor padrão Angular Material
+  font-weight: 500;
+
+  &:hover {
+    text-decoration: underline;
+  }
+
+  mat-icon {
+    font-size: 16px;
+    margin-left: 4px;
+  }
+}
